Now using float instead of int for delta.
The reasoning for this is that it is more difficult to turn int to float than
it is to turn float to int. Where the framerate is high then a float delta could
be as little as 0.0001 (seconds), and if you multiplied that by 1000 to get
milliseconds you end up with 0.1, and rounding that returns you 0.

But this is up for debate. If you wish to use hard int, you could access it
elsewhere than using the world instance. The delta float in world would only
be used in the interval and delayed systems, and frankly a slight imprecision
doesn't matter a whole lot there. You could make your own interval systems
based on the bundled ones, using int.

In the end I may throw deltas out of the World entirely so the framework is
delta neutral. I do not want it to dictate these things.

+++ src/com/artemis/systems/DelayedEntityProcessingSystem.java
@@ -14,10 +14,10 @@ public abstract class DelayedEntityProcessingSystem extends DelayedEntitySystem 
 	 * Process a entity this system is interested in.
 	 * @param e the entity to process.
 	 */
-	protected abstract void process(Entity e, int accumulatedDelta);
+	protected abstract void process(Entity e, float accumulatedDelta);
 
 	@Override
-	protected final void processEntities(ImmutableBag<Entity> entities, int accumulatedDelta) {
+	protected final void processEntities(ImmutableBag<Entity> entities, float accumulatedDelta) {
 		for (int i = 0, s = entities.size(); s > i; i++) {
 			process(entities.get(i), accumulatedDelta);
 		}
